backend/requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements-dev.txt requirements-dev.in
#
alabaster==1.0.0
    # via sphinx
asgiref==3.11.0
    # via daphne
astroid==2.15.8
    # via flake8-django
asttokens==3.0.0
    # via stack-data
attrs==25.4.0
    # via
    #   service-identity
    #   twisted
autobahn==25.12.2
    # via daphne
automat==25.4.16
    # via twisted
babel==2.17.0
    # via sphinx
black==25.1.0
    # via -r requirements-dev.in
build==1.3.0
    # via pip-tools
cbor2==5.8.0
    # via autobahn
certifi==2025.8.3
    # via requests
cffi==2.0.0
    # via
    #   autobahn
    #   cryptography
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via
    #   black
    #   pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
    #   ipython
    #   pytest
    #   sphinx
comm==0.2.3
    # via ipykernel
constantly==23.10.4
    # via twisted
cryptography==46.0.3
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
daphne==4.2.1
    # via -r requirements-dev.in
debugpy==1.8.16
    # via ipykernel
decorator==5.2.1
    # via ipython
distlib==0.4.0
    # via virtualenv
docutils==0.21.2
    # via sphinx
executing==2.2.1
    # via stack-data
factory-boy==3.3.3
    # via -r requirements-dev.in
Faker==40.1.2
    # via factory-boy
filelock==3.20.0
    # via virtualenv
flake8==6.1.0
    # via
    #   -r requirements-dev.in
    #   flake8-django
flake8-django==1.4
    # via -r requirements-dev.in
flake8-docstrings==1.7.0
    # via -r requirements-dev.in
flake8-bugbear==24.12.12
    # via -r requirements-dev.in
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
identify==2.6.15
    # via pre-commit
idna==3.10
    # via
    #   hyperlink
    #   requests
    #   twisted
imagesize==1.4.1
    # via sphinx
incremental==24.11.0
    # via twisted
iniconfig==2.1.0
    # via pytest
ipykernel==6.30.1
    # via -r requirements-dev.in
ipython==9.5.0
    # via
    #   -r requirements-dev.in
    #   ipykernel
ipython-pygments-lexers==1.1.1
    # via ipython
isort==7.0.0
    # via -r requirements-dev.in
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via sphinx
jupyter-client==8.6.3
    # via ipykernel
jupyter-core==5.8.1
    # via
    #   ipykernel
    #   jupyter-client
lazy-object-proxy==1.12.0
    # via astroid
markupsafe==3.0.2
    # via jinja2
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mccabe==0.7.0
    # via flake8
msgpack==1.1.2
    # via autobahn
mypy==1.17.1
    # via -r requirements-dev.in
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
nest-asyncio==1.6.0
    # via ipykernel
nodeenv==1.9.1
    # via pre-commit
packaging==25.0
    # via
    #   black
    #   build
    #   incremental
    #   ipykernel
    #   pytest
    #   sphinx
parso==0.8.5
    # via jedi
pathspec==0.12.1
    # via
    #   black
    #   mypy
pip-tools==7.5.0
    # via -r requirements-dev.in
platformdirs==4.4.0
    # via
    #   black
    #   jupyter-core
    #   virtualenv
pluggy==1.6.0
    # via pytest
pre-commit==4.3.0
    # via -r requirements-dev.in
prompt-toolkit==3.0.52
    # via ipython
psutil==7.0.0
    # via ipykernel
pure-eval==0.2.3
    # via stack-data
py-ubjson==0.16.1
    # via autobahn
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.2
    # via service-identity
pycodestyle==2.11.1
    # via flake8
pycparser==3.0
    # via cffi
pyflakes==3.1.0
    # via flake8
pydocstyle==6.3.0
    # via flake8-docstrings
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
    #   pytest
    #   sphinx
pyopenssl==25.3.0
    # via twisted
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.4.2
    # via
    #   -r requirements-dev.in
    #   pytest-asyncio
    #   pytest-django
pytest-asyncio==1.1.0
    # via -r requirements-dev.in
pytest-django==4.11.1
    # via -r requirements-dev.in
python-dateutil==2.9.0.post0
    # via jupyter-client
pywin32==311
    # via jupyter-core
pyyaml==6.0.3
    # via pre-commit
pyzmq==27.0.2
    # via
    #   ipykernel
    #   jupyter-client
requests==2.32.5
    # via sphinx
roman-numerals-py==3.1.0
    # via sphinx
service-identity==24.2.0
    # via twisted
six==1.17.0
    # via python-dateutil
snowballstemmer==3.0.1
    # via sphinx
sphinx==8.2.3
    # via -r requirements-dev.in
sphinxcontrib-applehelp==2.0.0
    # via sphinx
sphinxcontrib-devhelp==2.0.0
    # via sphinx
sphinxcontrib-htmlhelp==2.1.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==2.0.0
    # via sphinx
sphinxcontrib-serializinghtml==2.0.0
    # via sphinx
stack-data==0.6.3
    # via ipython
tornado==6.5.2
    # via
    #   ipykernel
    #   jupyter-client
traitlets==5.14.3
    # via
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
twisted[tls]==25.5.0
    # via daphne
txaio==25.12.2
    # via autobahn
typing-extensions==4.15.0
    # via
    #   mypy
    #   pyopenssl
    #   twisted
ujson==5.11.0
    # via autobahn
urllib3==2.5.0
    # via requests
virtualenv==20.35.3
    # via pre-commit
wcwidth==0.2.13
    # via prompt-toolkit
wheel==0.45.1
    # via pip-tools
wrapt==1.17.3
    # via astroid
zope-interface==8.2
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
